requirements.txt
aiogram==2.10.1
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
Babel==2.8.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cloudinary==1.22.0
cryptography==3.1.1
decorator==4.4.2
idna==2.10
multidict==4.7.6
pycparser==2.20
python-dotenv==0.14.0
python-telegram-bot==12.8
pytz==2020.1
six==1.15.0
tornado==6.0.4
urllib3==1.25.10
yarl==1.6.0
